fix(db): users 表主键 id 改为 INT 以对齐应用层整型主键
app/model 中 User.id 及 UserInfo.id 均为 int，建表 SQL 统一为 INT，
避免类型不一致导致查询结果映射异常。

+++ template/sqls/01_users.sql
@@ -1,6 +1,6 @@
 -- 用户表：开放注册、找回密码
 CREATE TABLE IF NOT EXISTS `users` (
-  `id`            BIGINT        NOT NULL AUTO_INCREMENT,
+  `id`            INT           NOT NULL AUTO_INCREMENT,
   `username`      VARCHAR(64)   NOT NULL,
   `email`         VARCHAR(255)  NOT NULL,
   `password_hash` VARCHAR(255)  NOT NULL,
